requirements.txt
# Core ML Libraries
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0

# Data Processing
pandas>=1.5.0
numpy>=1.24.0
opencv-python>=4.7.0
Pillow>=9.5.0

# Scientific Computing
scipy>=1.10.0
scikit-learn>=1.3.0

# Progress and Logging
tqdm>=4.65.0
wandb>=0.15.0
tensorboard>=2.13.0

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0

# HTTP Requests
requests>=2.31.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Optional: For teacher model (Qwen2.5-VL)
# qwen-vl-utils  # Uncomment if using Qwen teacher model
# qwen-vl  # Uncomment if using Qwen teacher model

# Optional: For OCR (Google Cloud Document AI)
# google-cloud-documentai>=2.20.0  # Uncomment if using Google Cloud OCR

# Optional: For document processing
# python-docx>=1.0.0  # Uncomment if needed

